mad/sched_opt now features a raw interface + a pack interface
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@13995 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/schedulers/sched_opt/src/nm_so_pkt_wrap.c
@@ -396,7 +396,7 @@ nm_so_pw_finalize(struct nm_so_pkt_wrap *p_so_pw)
       vec++;
       ptr = (void *)nm_so_aligned((uint64_t)vec->iov_base);
       remaining_bytes = vec->iov_len
-	                - ((uint64_t)ptr - (uint64_t)vec->iov_base);
+	                - (ptr - vec->iov_base);
       to_skip = 0;
     }
